AppStatus has now its own component
